+++ src/components/select/_select.base.ts
@@ -38,7 +38,10 @@ export default class StSelectBase extends Vue {
   suffixIcon!: string;
 
   @Prop({ type: Boolean, default: true })
-  clearIconAsSuffixIcon!: boolean;
+  showArrowIcon!: boolean;
+
+  @Prop({ type: Boolean, default: true })
+  clearIconAsArrowIcon!: boolean;
 
   @Prop({ type: Boolean, default: true })
   closeOnClear!: boolean;