+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.2.12
 
+### Features
+* `Select`'s prop `clearIconAsSuffixIcon` was renamed to `clearIconAsArrowIcon`. Also a new prop `showArrowIcon` was added.
+Now you can set a suffix-icon near the default arrow icon.
+
 ### Maintenance
 * Updated `Notification` component. Do not show browser notifications when window is active
 * Fixed `Select`'s `list` slot. Didn't work before